refactor: move derive-private-keys higher in the tree for reuse
